chore: cleanup console. re-org crates
